app/requirements.txt
Flask==0.12.2
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-HTTPAuth==3.2.2
Flask-SQLAlchemy==2.1
SQLAlchemy==1.1.10
alembic==0.9.2
ForgeryPy3==0.3.1
